requirements.txt
aiofiles==22.1.0
aiohttp==3.8.3
aiohttp-swagger==1.0.16
aiosignal==1.3.1
async-lru==1.0.3
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
cryptography==38.0.3
dill==0.3.6
distro==1.8.0
frozenlist==1.3.3
genie==20.10
genie.libs.clean==20.10.1
genie.libs.conf==20.10
genie.libs.filetransferutils==20.10
genie.libs.health==20.10.1
genie.libs.ops==20.10
genie.libs.parser==20.10
genie.libs.sdk==20.10
gitdb==4.0.10
GitPython==3.1.18
idna==3.4
Jinja2==3.1.2
jsonpickle==2.2.0
junit-xml==1.9
MarkupSafe==2.1.1
multidict==6.0.2
netaddr==0.8.0
packaging==21.3
pathspec==0.10.2
ply==3.11
prettytable==3.5.0
psutil==5.9.4
pyasn1==0.4.8
pyats==22.10
pyats.aereport==22.10
pyats.aetest==22.10
pyats.async==22.10
pyats.connections==22.10
pyats.datastructures==22.10
pyats.easypy==22.10
pyats.kleenex==22.10
pyats.log==22.10
pyats.reporter==22.10
pyats.results==22.10
pyats.tcl==22.10
pyats.topology==22.10
pyats.utils==22.10
pycparser==2.21
pycryptodomex==3.16.0
pyparsing==3.0.9
pysmi==0.3.4
pysnmp==4.4.12
python-engineio==3.14.2
python-socketio==4.6.1
PyYAML==6.0
requests==2.28.1
requests-toolbelt==0.9.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
smmap==5.0.0
tqdm==4.64.1
unicon==22.10
unicon.plugins==22.10
urllib3==1.26.13
virl2-client 
wcwidth==0.2.5
xmltodict==0.13.0
yamllint==1.28.0
yarl==1.8.1
